+++ apps/web/src/components/Merkl/MerklSection.tsx
@@ -39,10 +39,19 @@ function TextWarning({ tokenAmount }: { tokenAmount: CurrencyAmount<Currency> })
   )
 }
 
+const LearnMoreLink = () => {
+  const { t } = useTranslation()
+
+  return (
+    <Link color="currentColor" fontSize="md" external style={{ display: 'inline-flex' }} href="https://docs.merkl.xyz/">
+      {t('Learn more about Merkl')}
+    </Link>
+  )
+}
+
 export function MerklSection({
   poolAddress,
   notEnoughLiquidity,
-  isStakedInMCv3,
   outRange,
   disabled,
 }: {
@@ -50,25 +59,12 @@ export function MerklSection({
   disabled: boolean
   poolAddress: string | null
   notEnoughLiquidity: boolean
-  isStakedInMCv3: boolean
 }) {
   const { t } = useTranslation()
 
   const { claimTokenReward, isClaiming, rewardsPerToken, hasMerkl } = useMerkl(poolAddress)
 
   if (!rewardsPerToken.length || (!hasMerkl && rewardsPerToken.every((r) => r.equalTo('0')))) return null
-
-  const learnMoreComp = (
-    <Link
-      color="currentColor"
-      fontSize="md"
-      external
-      style={{ display: 'inline-flex' }}
-      href="https://docs.angle.money/merkl/introduction"
-    >
-      {t('Learn more about Merkl')}
-    </Link>
-  )
 
   return (
     <Column justifyContent="space-between" gap="8px" width="100%" ml={['0px', '0px', '16px', '16px']} mt="24px">
@@ -84,12 +80,7 @@ export function MerklSection({
           {isClaiming ? t('Claiming...') : t('Claim')}
         </Button>
       </AutoRow>
-      <LightGreyCard
-        mr="4px"
-        style={{
-          padding: '16px 8px',
-        }}
-      >
+      <LightGreyCard mr="4px" padding="16px 8px">
         {rewardsPerToken.map((tokenAmount) => (
           <AutoRow justifyContent="space-between">
             <Flex>
@@ -110,7 +101,7 @@ export function MerklSection({
           <MessageText color="textSubtle">
             {t('This Merkl campaign is NOT rewarding out-of-range liquidity. To earn rewards, adjust your position.')}
             <br />
-            {learnMoreComp}
+            <LearnMoreLink />
           </MessageText>
         </Message>
       ) : hasMerkl ? (
@@ -127,12 +118,12 @@ export function MerklSection({
               external
               color="currentColor"
               style={{ display: 'inline-flex' }}
-              href="https://merkl.angle.money/?times=active%2Cfuture%2C&phrase=PancakeSwap"
+              href="https://merkl.angle.money/?search=PancakeSwap&status=live%2Csoon"
             >
               {t('here')}
-            </Link>{' '}
+            </Link>
             <br />
-            {learnMoreComp}
+            <LearnMoreLink />
           </MessageText>
         </Message>
       ) : null}