+++ apps/web/src/hooks/useTokenComparator.ts
@@ -1,0 +1,14 @@
+import { Token, getTokenComparator } from '@pancakeswap/sdk'
+import { useMemo } from 'react'
+import { useAllTokenBalances } from '../state/wallet/hooks'
+
+export function useTokenComparator(inverted: boolean): (tokenA: Token, tokenB: Token) => number {
+  const balances = useAllTokenBalances()
+  const comparator = useMemo(() => getTokenComparator(balances ?? {}), [balances])
+  return useMemo(() => {
+    if (inverted) {
+      return (tokenA: Token, tokenB: Token) => comparator(tokenA, tokenB) * -1
+    }
+    return comparator
+  }, [inverted, comparator])
+}