+++ apps/web/src/utils/customGTMEventTracking.ts
@@ -2,16 +2,21 @@ export enum GTMEvent {
   EventTracking = 'eventTracking',
   Swap = 'swap',
   AddLiquidity = 'addLiquidity',
+  RemoveLiquidity = 'removeLiquidity',
   Farm = 'stakeFarm',
+  UnStakeFarm = 'unStakeFarm',
   WalletConnect = 'walletConnect',
   Web3WalletView = 'Web3WalletView',
+  MenuClick = 'menuClick',
 }
 
 export enum GTMCategory {
   TokenHighlight = 'TokenHighlight',
   Swap = 'Swap',
   AddLiquidity = 'AddLiquidity',
+  RemoveLiquidity = 'RemoveLiquidity',
   Farm = 'Farm',
+  UnStakeFarm = 'unStakeFarm',
   WalletConnect = 'WalletConnect',
   Web3WalletView = 'Web3WalletView',
 }
@@ -20,15 +25,17 @@ export enum GTMAction {
   ClickTradeButton = 'Click Trade Button',
   ClickSwapButton = 'Click Swap Button',
   ClickAddLiquidityButton = 'Click Add Liquidity Button',
+  ClickRemoveLiquidityButton = 'Click Remove Liquidity Button',
   ClickStakeButton = 'Click Stake Button',
+  ClickUnStakeButton = 'Click UnStake Button',
   ClickWalletConnectButton = 'Click Wallet Connect and Connected',
   Web3WalletView = 'Web3 Wallet Page View',
 }
 
 interface CustomGTMDataLayer {
   event: GTMEvent
-  category: GTMCategory
-  action: GTMAction
+  category?: GTMCategory
+  action?: GTMAction
   label?: string
 }
 
@@ -69,12 +76,30 @@ export const logGTMClickStakeFarmEvent = () => {
   })
 }
 
+export const logGTMClickUnStakeFarmEvent = () => {
+  console.info('---UnStake---')
+  window?.dataLayer?.push({
+    event: GTMEvent.UnStakeFarm,
+    action: GTMAction.ClickUnStakeButton,
+    category: GTMCategory.UnStakeFarm,
+  })
+}
+
 export const logGTMClickAddLiquidityEvent = () => {
   console.info('---AddLiquidity---')
   window?.dataLayer?.push({
     event: GTMEvent.AddLiquidity,
     action: GTMAction.ClickAddLiquidityButton,
     category: GTMCategory.AddLiquidity,
+  })
+}
+
+export const logGTMClickRemoveLiquidityEvent = () => {
+  console.info('---RemoveLiquidity---')
+  window?.dataLayer?.push({
+    event: GTMEvent.RemoveLiquidity,
+    action: GTMAction.ClickRemoveLiquidityButton,
+    category: GTMCategory.RemoveLiquidity,
   })
 }
 
@@ -94,5 +119,12 @@ export const logWeb3WalletViews = () => {
     event: GTMEvent.Web3WalletView,
     action: GTMAction.Web3WalletView,
     category: GTMCategory.Web3WalletView,
+  })
+}
+
+export const logMenuClick = (path: string) => {
+  window?.dataLayer?.push({
+    event: GTMEvent.MenuClick,
+    label: path,
   })
 }