+++ apps/web/src/views/Farms/components/FarmCard/V3/FarmV3ApyButton.tsx
@@ -5,6 +5,8 @@ import {
   CalculateIcon,
   Flex,
   IconButton,
+  LinkExternal,
+  PairDataTimeWindowEnum,
   RocketIcon,
   Skeleton,
   Text,
@@ -22,18 +24,18 @@ import { useCakePrice } from 'hooks/useCakePrice'
 import { useContext, useMemo, useState } from 'react'
 import { styled } from 'styled-components'
 
+import isUndefinedOrNull from '@pancakeswap/utils/isUndefinedOrNull'
 import { Bound } from 'config/constants/types'
 import { usePoolAvgInfo } from 'hooks/usePoolAvgInfo'
 import { usePairTokensPrice } from 'hooks/v3/usePairTokensPrice'
 import { useAllV3Ticks } from 'hooks/v3/usePoolTickData'
 import useV3DerivedInfo from 'hooks/v3/useV3DerivedInfo'
+import { type V3Farm } from 'state/farms/types'
 import { useFarmsV3Public } from 'state/farmsV3/hooks'
 import { Field } from 'state/mint/actions'
 import LiquidityFormProvider from 'views/AddLiquidityV3/formViews/V3FormView/form/LiquidityFormProvider'
 import { useV3FormState } from 'views/AddLiquidityV3/formViews/V3FormView/form/reducer'
-import { V3Farm } from 'views/Farms/FarmsV3'
 import { FarmsV3Context } from 'views/Farms'
-import isUndefinedOrNull from '@pancakeswap/utils/isUndefinedOrNull'
 import { USER_ESTIMATED_MULTIPLIER, useUserPositionInfo } from '../../YieldBooster/hooks/bCakeV3/useBCakeV3Info'
 import { BoostStatus, useBoostStatus } from '../../YieldBooster/hooks/bCakeV3/useBoostStatus'
 import { getDisplayApr } from '../../getDisplayApr'
@@ -45,11 +47,35 @@ const ApyLabelContainer = styled(Flex)`
   }
 `
 
+const StyledLi = styled.li`
+  flex-wrap: nowrap;
+  display: flex;
+  gap: 5px;
+  position: relative;
+  padding-left: 22px;
+  &::before {
+    content: '';
+    position: absolute;
+    transform: translateY(-50%);
+    top: 50%;
+    border-radius: 50%;
+    left: 0;
+    width: 6px;
+    height: 6px;
+    background: ${({ theme }) => theme.colors.text};
+  }
+`
+
 type FarmV3ApyButtonProps = {
   farm: V3Farm
   existingPosition?: Position
   isPositionStaked?: boolean
   tokenId?: string
+  additionAprInfo?: {
+    aprValue: number
+    aprTitle: string
+    aprLink: string
+  }
 }
 
 export function FarmV3ApyButton(props: FarmV3ApyButtonProps) {
@@ -60,13 +86,19 @@ export function FarmV3ApyButton(props: FarmV3ApyButtonProps) {
   )
 }
 
-function FarmV3ApyButton_({ farm, existingPosition, isPositionStaked, tokenId }: FarmV3ApyButtonProps) {
+function FarmV3ApyButton_({
+  farm,
+  existingPosition,
+  isPositionStaked,
+  tokenId,
+  additionAprInfo,
+}: FarmV3ApyButtonProps) {
   const { farmsAvgInfo } = useContext(FarmsV3Context)
   const { token: baseCurrency, quoteToken: quoteCurrency, feeAmount, lpAddress } = farm
   const { t } = useTranslation()
   const roiModal = useModalV2()
 
-  const [priceTimeWindow, setPriceTimeWindow] = useState(0)
+  const [priceTimeWindow, setPriceTimeWindow] = useState(PairDataTimeWindowEnum.DAY)
   const prices = usePairTokensPrice(lpAddress, priceTimeWindow, baseCurrency?.chainId, roiModal.isOpen)
 
   const { ticks: data } = useAllV3Ticks(baseCurrency, quoteCurrency, feeAmount, roiModal.isOpen)
@@ -181,20 +213,25 @@ function FarmV3ApyButton_({ farm, existingPosition, isPositionStaked, tokenId }:
   const lpApr = existingPosition ? +apr.toFixed(2) : globalLpApr
   const cakeApr = +(farm.cakeApr ?? 0)
 
-  const displayApr = getDisplayApr(cakeApr, lpApr)
+  const displayApr = getDisplayApr(cakeApr, lpApr, additionAprInfo?.aprValue)
   const cakeAprDisplay = cakeApr.toFixed(2)
   const positionCakeAprDisplay = positionCakeApr.toFixed(2)
   const lpAprDisplay = lpApr.toFixed(2)
+  const additionalAprDisplay = (additionAprInfo?.aprValue ?? 0).toFixed(2)
   const { isDesktop } = useMatchBreakpoints()
   const {
     data: { boostMultiplier },
   } = useUserPositionInfo(tokenId ?? '-1')
 
   const estimatedAPR = useMemo(() => {
-    return (parseFloat(cakeAprDisplay) * USER_ESTIMATED_MULTIPLIER + parseFloat(lpAprDisplay)).toLocaleString('en-US', {
+    return (
+      parseFloat(cakeAprDisplay) * USER_ESTIMATED_MULTIPLIER +
+      parseFloat(lpAprDisplay) +
+      (additionAprInfo?.aprValue ?? 0)
+    ).toLocaleString('en-US', {
       maximumFractionDigits: 2,
     })
-  }, [cakeAprDisplay, lpAprDisplay])
+  }, [cakeAprDisplay, lpAprDisplay, additionAprInfo])
   const canBoosted = useMemo(() => boostedStatus !== BoostStatus.CanNotBoost, [boostedStatus])
   const isBoosted = useMemo(() => boostedStatus === BoostStatus.Boosted, [boostedStatus])
   const positionDisplayApr = getDisplayApr(+positionCakeApr, lpApr)
@@ -221,6 +258,16 @@ function FarmV3ApyButton_({ farm, existingPosition, isPositionStaked, tokenId }:
         <li>
           {t('LP Fee APR')}: <b>{lpAprDisplay}%</b>
         </li>
+        {additionAprInfo && (
+          <StyledLi>
+            <Flex style={{ flexWrap: 'nowrap', alignItems: 'center', gap: 5 }}>
+              {additionAprInfo.aprTitle}: <b>{additionalAprDisplay}%</b>
+              <LinkExternal display="inline-block" href={additionAprInfo.aprLink}>
+                {t('Check')}
+              </LinkExternal>
+            </Flex>
+          </StyledLi>
+        )}
       </ul>
       <br />
       <Text>
@@ -256,6 +303,14 @@ function FarmV3ApyButton_({ farm, existingPosition, isPositionStaked, tokenId }:
         <li>
           {t('LP Fee APR')}: <b>{lpAprDisplay}%</b>
         </li>
+        {additionAprInfo && (
+          <StyledLi style={{ whiteSpace: 'nowrap', display: 'flex', flexWrap: 'nowrap', alignItems: 'center', gap: 5 }}>
+            {additionAprInfo.aprTitle}: <b>{additionalAprDisplay}%</b>
+            <LinkExternal display="inline-block" href={additionAprInfo.aprLink}>
+              {t('Check')}
+            </LinkExternal>
+          </StyledLi>
+        )}
       </ul>
     </>,
   )
@@ -361,6 +416,7 @@ function FarmV3ApyButton_({ farm, existingPosition, isPositionStaked, tokenId }:
           prices={prices}
           priceSpan={priceTimeWindow}
           onPriceSpanChange={setPriceTimeWindow}
+          additionalApr={additionAprInfo?.aprValue}
         />
       )}
     </>