+++ apps/web/src/views/Farms/components/FarmTable/Row.tsx
@@ -6,10 +6,10 @@ import { FarmWidget } from '@pancakeswap/widgets-internal'
 import { createElement, useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { styled } from 'styled-components'
 
-import { useActiveChainId } from 'hooks/useActiveChainId'
 import { useMerklInfo } from 'hooks/useMerkl'
+import { type V3Farm } from 'state/farms/types'
 import { useMerklUserLink } from 'utils/getMerklLink'
-import { V2Farm, V3Farm } from 'views/Farms/FarmsV3'
+import { V2Farm } from 'views/Farms/FarmsV3'
 import { useBCakeBoostLimitAndLockInfo } from 'views/Farms/components/YieldBooster/hooks/bCakeV3/useBCakeV3Info'
 import { RewardPerDay } from 'views/PositionManagers/components/RewardPerDay'
 import { FarmV3ApyButton } from '../FarmCard/V3/FarmV3ApyButton'
@@ -122,7 +122,6 @@ const Row: React.FunctionComponent<React.PropsWithChildren<RowPropsWithLoading>>
   const { tokenIds } = useUserBoostedPoolsTokenId()
   const { isBoosted } = useIsSomePositionBoosted(props.type === 'v3' ? props?.details?.stakedPositions : [], tokenIds)
   const { locked } = useBCakeBoostLimitAndLockInfo()
-  const { chainId } = useActiveChainId()
   const toggleActionPanel = useCallback(() => {
     setActionPanelExpanded(!actionPanelExpanded)
   }, [actionPanelExpanded])
@@ -148,6 +147,7 @@ const Row: React.FunctionComponent<React.PropsWithChildren<RowPropsWithLoading>>
   const merklUserLink = useMerklUserLink()
 
   const { merklApr } = useMerklInfo(farm?.merklLink ? props.details.lpAddress : null)
+
   return (
     <>
       {!isMobile ? (
@@ -202,7 +202,14 @@ const Row: React.FunctionComponent<React.PropsWithChildren<RowPropsWithLoading>>
                     <td key={key}>
                       <CellInner>
                         <CellLayout label={t('APR')}>
-                          <FarmV3ApyButton farm={props.details} />
+                          <FarmV3ApyButton
+                            farm={props.details}
+                            additionAprInfo={
+                              merklApr && farm.merklLink
+                                ? { aprTitle: t('Merkl APR'), aprValue: merklApr, aprLink: farm.merklLink }
+                                : undefined
+                            }
+                          />
                         </CellLayout>
                       </CellInner>
                     </td>
@@ -348,7 +355,14 @@ const Row: React.FunctionComponent<React.PropsWithChildren<RowPropsWithLoading>>
               <AprMobileCell>
                 <CellLayout label={t('APR')}>
                   {props.type === 'v3' ? (
-                    <FarmV3ApyButton farm={props.details} />
+                    <FarmV3ApyButton
+                      farm={props.details}
+                      additionAprInfo={
+                        merklApr && farm.merklLink
+                          ? { aprTitle: t('Merkl APR'), aprValue: merklApr, aprLink: farm.merklLink }
+                          : undefined
+                      }
+                    />
                   ) : (
                     <>
                       <Apr