+++ apps/web/src/views/V3Info/types.ts
@@ -33,6 +33,28 @@ export enum TransactionType {
   BURN,
 }
 
+export type ApiTransaction = {
+  type: 'mint' | 'burn' | 'swap'
+  transactionHash: string
+  timestamp: Record<string, never> | string
+  origin: null | string
+  token0: {
+    id: string
+    symbol: string
+    name: string
+    decimals: number
+  }
+  token1: {
+    id: string
+    symbol: string
+    name: string
+    decimals: number
+  }
+  amountUSD: string
+  amount0: string
+  amount1: string
+}
+
 export type Transaction = {
   type: TransactionType
   hash: string
@@ -95,6 +117,34 @@ export type PoolChartEntry = {
   volumeUSD: number
   totalValueLockedUSD: number
   feesUSD: number
+}
+
+export interface ApiPoolData {
+  feeTier: number
+  id: string
+  volumeUSD24h: string
+  volumeUSD7d: string
+  token0: {
+    id: string
+    symbol: string
+    name: string
+    decimals: number
+  }
+  token1: {
+    id: string
+    symbol: string
+    name: string
+    decimals: number
+  }
+  totalFeeUSD: any
+  liquidity: string
+  sqrtPrice: string
+  tick: number | null
+  tvlUSD: string
+  token0Price: string
+  token1Price: string
+  tvlToken0: string
+  tvlToken1: string
 }
 
 export interface PoolData {