+++ packages/widgets-internal/components/Banner/BannerTitle.tsx
@@ -2,10 +2,10 @@ import { Text, TextProps, breakpoints } from "@pancakeswap/uikit";
 import { PropsWithChildren } from "react";
 import styled from "styled-components";
 
-type VariantTypes = "purple" | "orange" | "yellow" | "listaBlue";
+type VariantTypes = "purple" | "orange" | "yellow" | "listaBlue" | "green";
 
 type Props = {
-  variant: VariantTypes;
+  variant: VariantTypes | Variant;
   fontSize?: number;
   lineHeight?: number;
   strokeColor?: string;
@@ -52,6 +52,14 @@ const variants: { [key in VariantTypes]: Variant } = {
     strokeSize: 0,
     fontSize: 40,
     lineHeight: 44,
+    fontWeight: 800,
+  },
+  green: {
+    color: "#50E892",
+    strokeColor: "#50E892",
+    strokeSize: 0,
+    fontSize: 28,
+    lineHeight: 31,
     fontWeight: 800,
   },
 };
@@ -127,7 +135,8 @@ export function BannerTitle({
   color: defaultColor,
   ...props
 }: PropsWithChildren<Omit<TextProps, "fontSize" | "lineHeight"> & Props>) {
-  const { color, fontSize, fontWeight, strokeSize, strokeColor, lineHeight } = variants[variant];
+  const variantData = typeof variant === "string" ? variants[variant] : variant;
+  const { color, fontSize, fontWeight, strokeSize, strokeColor, lineHeight } = variantData;
   return (
     <FancyText
       color={color ?? defaultColor}