+++ apps/games/CHANGELOG.md
@@ -1,5 +1,26 @@
 # games
 
+## 0.1.17
+
+### Patch Changes
+
+- @pancakeswap/hooks@0.0.32
+- @pancakeswap/utils@6.0.9
+- @pancakeswap/uikit@0.67.1
+
+## 0.1.16
+
+### Patch Changes
+
+- Updated dependencies [4e3ed54]
+- Updated dependencies [6eac78f]
+- Updated dependencies [76f049e]
+- Updated dependencies [c9c6c44]
+- Updated dependencies [57bac78]
+- Updated dependencies [42dfc77]
+- Updated dependencies [a3ca679]
+  - @pancakeswap/uikit@0.67.0
+
 ## 0.1.15
 
 ### Patch Changes