+++ packages/smart-router/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @pancakeswap/smart-router
 
+## 6.0.17
+
+### Patch Changes
+
+- Updated dependencies [f551e5e]
+  - @pancakeswap/swap-sdk-core@1.2.0
+  - @pancakeswap/stable-swap-sdk@1.0.15
+  - @pancakeswap/sdk@5.8.8
+  - @pancakeswap/token-lists@0.0.12
+  - @pancakeswap/v3-sdk@3.8.11
+  - @pancakeswap/gauges@1.2.13
+  - @pancakeswap/multicall@3.5.8
+  - @pancakeswap/tokens@0.6.23
+
+## 6.0.16
+
+### Patch Changes
+
+- Updated dependencies [911d107]
+  - @pancakeswap/v3-sdk@3.8.10
+  - @pancakeswap/gauges@1.2.12
+
+## 6.0.15
+
+### Patch Changes
+
+- Updated dependencies [fe96bb1]
+  - @pancakeswap/v3-sdk@3.8.9
+  - @pancakeswap/gauges@1.2.11
+
+## 6.0.14
+
+### Patch Changes
+
+- Updated dependencies [4628ab7]
+- Updated dependencies [8f573cf]
+  - @pancakeswap/tokens@0.6.22
+  - @pancakeswap/gauges@1.2.10
+  - @pancakeswap/stable-swap-sdk@1.0.14
+
 ## 6.0.13
 
 ### Patch Changes