+++ packages/universal-router-sdk/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @pancakeswap/universal-router-sdk
 
+## 1.0.30
+
+### Patch Changes
+
+- @pancakeswap/smart-router@6.0.17
+- @pancakeswap/sdk@5.8.8
+- @pancakeswap/v3-sdk@3.8.11
+- @pancakeswap/permit2-sdk@1.0.10
+
+## 1.0.29
+
+### Patch Changes
+
+- Updated dependencies [911d107]
+  - @pancakeswap/v3-sdk@3.8.10
+  - @pancakeswap/smart-router@6.0.16
+
+## 1.0.28
+
+### Patch Changes
+
+- Updated dependencies [fe96bb1]
+  - @pancakeswap/v3-sdk@3.8.9
+  - @pancakeswap/smart-router@6.0.15
+
+## 1.0.27
+
+### Patch Changes
+
+- @pancakeswap/smart-router@6.0.14
+
 ## 1.0.26
 
 ### Patch Changes